requirements.txt
numpy
matplotlib
Flask
Pillow
gunicorn
flask_cors
scikit-learn